Replace adminApiProvider with reactAdminDataProvider and remove unused files
