+++ src/module/Ingredients/index.ts
@@ -1,2 +1,2 @@
 export { IngredientList, IngredientEdit, IngredientCreate } from './ui'
-export { ingredientsDataProvider } from './model'
+export { ingredientsDataProvider, type CreateIngredientData } from './model'
